requirements/base.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Caching==2.1.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
redis==5.0.1
bcrypt==4.0.1
python-dotenv==1.0.0
celery==5.3.4
gunicorn==21.2.0
psycopg2-binary==2.9.7
requests==2.31.0
Babel==2.13.1
APScheduler==3.10.4